+++ app/commands/__init__.py
@@ -1,22 +1,27 @@
-from app.commands import CommandHandler
-from app.commands import DiscordCommand
-from app.commands import ExitCommand
-from app.commands import GoodbyeCommand
-from app.commands import GreetCommand
-from app.commands import MenuCommand
+from abc import ABC, abstractmethod
+class Command(ABC):
+    @abstractmethod
+    def execute(self):
+        pass
 
-class App:
+class CommandHandler:
     def __init__(self):
-        self.command_handler = CommandHandler()
+        self.commands = {} #empty dictionary for commands to be stored
 
-    def start(self):
-        self.command_handler.register_command("greet", GreetCommand())
-        self.command_handler.register_command("goodbye", GoodbyeCommand())
-        self.command_handler.register_command("menu", MenuCommand())
-        self.command_handler.register_command("exit", ExitCommand())
-        self.command_handler.register_command("discord", DiscordCommand())
+    def register_command(self, command_name:str, command:Command):
+        self.commands[command_name] = command
 
-        print("Type 'Exit to quit the program")
-        while True:
-            self.command_handler.execute_command(input(">>> ").strip())
+    """def execute_command(self, command_name:str):
+         Look Before you Leap (LBYL) use when you expect the key to be missing
+        if command_name in self.commands:
+            self.commands[command_name].execute()
+        else:
+            print(f"Command '{command_name}' not found") """
+
+    """ Easier to ask for forgiveness than permission (EAFP) use when you expect the key to be present"""
+    def execute_command(self, command_name:str):
+        try:
+            self.commands[command_name].execute()
+        except KeyError:
+            print(f"Command '{command_name}' not found")
